chore(theme): Move to own folder
