+++ components/__tests__/AddFundsSourcePicker.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import 'jest-styled-components';
 
-import theme from '../../lib/constants/theme';
+import theme from '../../lib/theme';
 import { MockAddFundsSourcePicker, MockAddFundsSourcePickerForUser } from '../AddFundsSourcePicker';
 import { snapshotI18n } from '../../test/snapshot-helpers';
