+++ lib/theme/index.js
