+++ pages/_app.js
@@ -16,7 +16,7 @@ import UserProvider from '../components/UserProvider';
 import StripeProviderSSR from '../components/StripeProvider';
 import withData from '../lib/withData';
 
-import theme from '../lib/constants/theme';
+import theme from '../lib/theme';
 
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css'; // eslint-disable-line node/no-unpublished-import
 import '../node_modules/nprogress/nprogress.css'; // eslint-disable-line node/no-unpublished-import
