+++ styleguide/Wrapper.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { IntlProvider } from 'react-intl';
 
 import StripeProvider from '../components/StripeProvider';
-import theme from '../lib/constants/theme';
+import theme from '../lib/theme';
 
 const STRIPE_KEY = process.env.STRIPE_KEY || 'pk_test_5aBB887rPuzvWzbdRiSzV3QB';
